refactored test
